Back End/geo-info-provider/requirements.txt
aniso8601==8.0.0
attrs==20.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
flask-restplus==0.13.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
pyrsistent==0.17.0
pytz==2020.1
requests==2.25.1
six==1.15.0
urllib3==1.26.4
Werkzeug==0.16.1
zipp==3.1.0
